pyproject.toml
[project]
name = "egress-checker"
version = "0.1.0"
description = "Lightweight egress connectivity checker for AKS tenants"
readme = "README.md"
authors = [
    { name = "dudick123", email = "brian.dudick@outlook.com" }
]
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "prometheus-client>=0.24.1",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.15.5",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
]

[tool.coverage.run]
source = ["egress_checker"]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
line-length = 99
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "C4", "PT", "RUF"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
